@@ -1,12 +1,11 @@
 package org.reactome.server.tools.event.exporter;
 
 import com.martiansoftware.jsap.*;
-import org.reactome.server.graph.domain.model.TopLevelPathway;
 import org.reactome.server.graph.exception.CustomQueryException;
 import org.reactome.server.graph.service.*;
 import org.reactome.server.graph.service.util.DatabaseObjectUtils;
 import org.reactome.server.graph.utils.ReactomeGraphCore;
-import org.reactome.server.tools.event.exporter.config.ReactomeNeo4jConfig;
+import org.reactome.server.tools.event.exporter.config.EventPdfNeo4jConfig;
 import org.reactome.server.tools.event.exporter.util.ProgressBar;
 import org.slf4j.LoggerFactory;
 
@@ -26,16 +25,15 @@ public class Main {
 						new FlaggedOption("output", JSAP.STRING_PARSER, null, JSAP.REQUIRED, 'o', "output", "The output folder"),
 
 						// diagram options
-						new FlaggedOption("ehlds", JSAP.STRING_PARSER, null, JSAP.REQUIRED, 'e', "ehld", "The folder containing the EHLD svg files"),
-						new FlaggedOption("diagrams", JSAP.STRING_PARSER, null, JSAP.REQUIRED, 'd', "diagram", "The folder containing the diagram json files"),
+						new FlaggedOption("ehlds", JSAP.STRING_PARSER, null, JSAP.NOT_REQUIRED, 'e', "ehld", "The folder containing the EHLD svg files"),
+						new FlaggedOption("diagrams", JSAP.STRING_PARSER, null, JSAP.NOT_REQUIRED, 'd', "diagram", "The folder containing the diagram json files"),
 						new FlaggedOption("analysis", JSAP.STRING_PARSER, null, JSAP.NOT_REQUIRED, 'a', "analysis", "The folder containing the analysis files"),
-						new FlaggedOption("summary", JSAP.STRING_PARSER, null, JSAP.REQUIRED, 's', "summary", "The file containing the summary of pathways with EHLD assigned"),
+						new FlaggedOption("summary", JSAP.STRING_PARSER, null, JSAP.NOT_REQUIRED, 's', "summary", "The file containing the summary of pathways with EHLD assigned"),
 
 						// GRAPH-DB options
-						new FlaggedOption("host", JSAP.STRING_PARSER, "localhost", JSAP.NOT_REQUIRED, 'h', "host", "The neo4j host"),
-						new FlaggedOption("port", JSAP.STRING_PARSER, "7474", JSAP.NOT_REQUIRED, 'p', "port", "The neo4j port"),
+						new FlaggedOption("host", JSAP.STRING_PARSER, "bolt://localhost:7687", JSAP.NOT_REQUIRED, 'h', "host", "The neo4j host"),
 						new FlaggedOption("user", JSAP.STRING_PARSER, "neo4j", JSAP.NOT_REQUIRED, 'u', "user", "The neo4j user"),
-						new FlaggedOption("password", JSAP.STRING_PARSER, null, JSAP.REQUIRED, 'w', "password", "The neo4j password"),
+						new FlaggedOption("password", JSAP.STRING_PARSER, "neo4j", JSAP.NOT_REQUIRED, 'w', "password", "The neo4j password"),
 
 						new FlaggedOption("profile", JSAP.STRING_PARSER, "Modern", JSAP.NOT_REQUIRED, 'c', "profile", "The colour diagram [Modern or Standard]"),
 
@@ -51,10 +49,9 @@ public class Main {
 		//Initialising ReactomeCore Neo4j configuration
 		ReactomeGraphCore.initialise(
 				config.getString("host"),
-				config.getString("port"),
 				config.getString("user"),
 				config.getString("password"),
-				ReactomeNeo4jConfig.class
+				EventPdfNeo4jConfig.class
 		);
 
 		final String diagrams = config.getString("diagrams");
@@ -67,7 +64,7 @@ public class Main {
 			System.exit(1);
 		}
 
-		final Collection<TopLevelPathway> targets = getTargets(config.getStringArray("target"));
+		final Collection<SimplePathwayResult> targets = getTargets(config.getStringArray("target"));
 
 		final DiagramService ds = ReactomeGraphCore.getService(DiagramService.class);
 		final AdvancedDatabaseObjectService ados = ReactomeGraphCore.getService(AdvancedDatabaseObjectService.class);
@@ -81,7 +78,7 @@ public class Main {
 		}
 		int i = 0;
 		final ProgressBar bar = new ProgressBar();
-		for (TopLevelPathway target : targets) {
+		for (SimplePathwayResult target : targets) {
 			if (verbose) {
 				final double progress = (double) i / targets.size();
 				final String message = String.format("%3d / %-3d\t%-15s\t%s", i + 1, targets.size(), target.getStId(), target.getDisplayName());
@@ -100,7 +97,7 @@ public class Main {
 		bar.clear();
 	}
 
-	private static Collection<TopLevelPathway> getTargets(String[] target) {
+	private static Collection<SimplePathwayResult> getTargets(String[] target) {
 		AdvancedDatabaseObjectService advancedDatabaseObjectService = ReactomeGraphCore.getService(AdvancedDatabaseObjectService.class);
 		String query;
 		Map<String, Object> parametersMap = new HashMap<>();
@@ -115,10 +112,11 @@ public class Main {
 					dbIds.add(Long.parseLong(id));
 				}
 			}
+
 			query = "MATCH (p:TopLevelPathway) " +
-					"WHERE p.dbId IN {dbIds} OR p.stId IN {stIds} " +
+					"WHERE p.dbId IN $dbIds OR p.stId IN $stIds " +
 					"WITH DISTINCT p " +
-					"RETURN p " +
+					"RETURN p.stId as stId, p.displayName as displayName " +
 					"ORDER BY p.dbId";
 			parametersMap.put("dbIds", dbIds);
 			parametersMap.put("stIds", stIds);
@@ -127,32 +125,31 @@ public class Main {
 			if (aux.toLowerCase().equals("all")) {
 				query = "MATCH (p:TopLevelPathway)-[:species]->(s:Species) " +
 						"WITH DISTINCT p, s " +
-						"RETURN p " +
+						"RETURN p.stId as stId, p.displayName as displayName " +
 						"ORDER BY s.dbId, p.dbId";
 			} else if (DatabaseObjectUtils.isStId(aux)) {
-				query = "MATCH (p:TopLevelPathway{stId:{stId}}) RETURN DISTINCT p";
+				query = "MATCH (p:TopLevelPathway{stId:$stId}) RETURN DISTINCT p.stId as stId, p.displayName as displayName ";
 				parametersMap.put("stId", DatabaseObjectUtils.getIdentifier(aux));
 			} else if (DatabaseObjectUtils.isDbId(aux)) {
-				query = "MATCH (p:TopLevelPathway{dbId:{dbId}}) RETURN DISTINCT p";
+				query = "MATCH (p:TopLevelPathway{dbId:$dbId}) RETURN DISTINCT p.stId as stId, p.displayName as displayName ";
 				parametersMap.put("dbId", DatabaseObjectUtils.getIdentifier(aux));
 			} else {
 				if (verbose) System.out.println(String.format("Detected species '%s'", aux));
-				query = "MATCH (p:TopLevelPathway{speciesName:{speciesName}}) " +
+				query = "MATCH (p:TopLevelPathway{speciesName:$speciesName}) " +
 						"WITH DISTINCT p " +
-						"RETURN p " +
+						"RETURN p.stId as stId, p.displayName as displayName " +
 						"ORDER BY p.dbId";
 				parametersMap.put("speciesName", aux);
 			}
 		}
 
-		Collection<TopLevelPathway> pathways = new ArrayList<>();
+		Collection<SimplePathwayResult> pathways = new ArrayList<>();
 		try {
-			pathways = advancedDatabaseObjectService.getCustomQueryResults(TopLevelPathway.class, query, parametersMap);
+			pathways = advancedDatabaseObjectService.getCustomQueryResults(SimplePathwayResult.class, query, parametersMap);
 		} catch (CustomQueryException e) {
 			LoggerFactory.getLogger("document-exporter").error("Problem retrieving the target pathways", e);
 		}
 		return pathways;
-
 	}
 
 }
